… lib/std/crypto/aes.zig
… lib/std/crypto/aes/soft.zig