* Reorganize crypto/aes in order to separate parameters, implementations and modes.
* Add a zero-cost abstraction over the internal representation of a block, so that blocks can be kept in vector registers in optimized implementations.
* Add architecture-independent aesenc/aesdec/aesenclast/aesdeclast operations, so that any AES-based primitive can be implemented, including these that don't use the original key schedule (AES-PRF, AEGIS, Meowhash...)
* Add support for parallelization/wide blocks to take advantage of hardware implementations.
* Align T-tables to cache lines in the software implementations to slightly reduce side channels.
* Add an optimized implementation for modern Intel CPUs with AES-NI.
* Add new tests.
* Reimplement the counter mode to work with any block cipher, any endianness and to take advantage of wide blocks.
* Add benchmarks for AES.


This includes changes and suggestions from #6327.

Our current software AES implementation has side channels, that are especially concerning on microcontrollers. The new wide block API will mitigate the performance impact of a side channel resistant implementation (bitsliced/fixsliced).

The AES-NI optimized implementation removes these side channels and is more than 12 times faster.

Previous implementation:

```text
    aes128-single:   21055153 ops/s
    aes256-single:   15770629 ops/s
         aes128-8:   22323772 ops/s
         aes256-8:   16574030 ops/s
```

Optimized implementation:

```text
    aes128-single:   94198135 ops/s
    aes256-single:   70740096 ops/s
         aes128-8:  276163380 ops/s
         aes256-8:  209833137 ops/s
```