@jedisct1 2020-09-23
And now that the [AES fixslicing](https://eprint.iacr.org/2020/1123.pdf) paper is public, this is something we may want for our software implementation.

It requires a different internal representation for blocks, and blocks have to be processed in pairs. The changes introduced here prepare for this.